apps/extract-service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
motor==3.3.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
email-validator
aiosmtplib 
aiodns